# docs/openapi/auth-service.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Auth Service — Catálogo de Filmes",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "400": {
            "description": "E-mail já cadastrado",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Email já cadastrado"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "401": {
            "description": "E-mail ou senha não batem com nenhum usuário cadastrado",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Email ou senha inválidos"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioOut"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Credenciais inválidas ou expiradas"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "password-reset"
        ],
        "summary": "Forgot Password",
        "operationId": "forgot_password_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Forgot Password Auth Forgot Password Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "password-reset"
        ],
        "summary": "Reset Password",
        "operationId": "reset_password_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Reset Password Auth Reset Password Post"
                }
              }
            }
          },
          "400": {
            "description": "Token inexistente, expirado ou já usado",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Link inválido, expirado ou já utilizado"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Listar Usuarios",
        "description": "Só admin: lista todos os usuários cadastrados — é o que permite a\nação exclusiva de admin desta atividade (promover/rebaixar o papel de\nalguém, olhando essa lista pra decidir quem).",
        "operationId": "listar_usuarios_auth_admin_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsuarioOut"
                  },
                  "type": "array",
                  "title": "Response Listar Usuarios Auth Admin Users Get"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Credenciais inválidas ou expiradas"
                }
              }
            }
          },
          "403": {
            "description": "Autenticado, mas sem papel admin",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Acesso restrito a admins"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/admin/users/{usuario_id}/role": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Alterar Papel Usuario",
        "description": "Só admin: promove ou rebaixa o papel de qualquer usuário. Essa é a\nação exclusiva de admin de verdade — recusa com 403 quem não for\nadmin, mesmo chamando o endpoint direto (dependência require_admin).",
        "operationId": "alterar_papel_usuario_auth_admin_users__usuario_id__role_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Usuario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioOut"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Credenciais inválidas ou expiradas"
                }
              }
            }
          },
          "403": {
            "description": "Autenticado, mas sem papel admin",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Acesso restrito a admins"
                }
              }
            }
          },
          "400": {
            "description": "`role` enviado não é um dos papéis válidos",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Papel inválido. Use um de: cinefilo, nerd, stalker_do_tomhanks, admin"
                }
              }
            }
          },
          "404": {
            "description": "usuario_id não existe",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Usuário não encontrado"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ForgotPasswordIn": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ForgotPasswordIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginIn": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "title": "Senha"
          }
        },
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "title": "LoginIn"
      },
      "RegisterIn": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "maxLength": 72,
            "minLength": 8,
            "title": "Senha"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "email",
          "senha"
        ],
        "title": "RegisterIn"
      },
      "ResetPasswordIn": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "nova_senha": {
            "type": "string",
            "maxLength": 72,
            "minLength": 8,
            "title": "Nova Senha"
          }
        },
        "type": "object",
        "required": [
          "token",
          "nova_senha"
        ],
        "title": "ResetPasswordIn"
      },
      "RoleUpdateIn": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "RoleUpdateIn"
      },
      "TokenOut": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenOut"
      },
      "UsuarioOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "email",
          "role",
          "criado_em"
        ],
        "title": "UsuarioOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}